fix warnings on Mac in src/pdf
Fix these class of warnings:
- unused functions
- unused locals
- sign mismatch
- missing function prototypes
- missing newline at end of file
- 64 to 32 bit truncation

The changes prefer to link in dead code in the debug build
with 'if (false)' than to comment it out, but trivial cases
are commented out or sometimes deleted if it appears to be
a copy/paste error.
Review URL: https://codereview.appspot.com/6299047

git-svn-id: http://skia.googlecode.com/svn/trunk@4178 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/pdf/SkPDFCatalog.cpp
@@ -190,7 +190,7 @@ off_t SkPDFCatalog::setSubstituteResourcesOffsets(off_t fileOffset,
     SkTDArray<SkPDFObject*>* targetList = getSubstituteList(firstPage);
     off_t offsetSum = fileOffset;
     for (int i = 0; i < targetList->count(); ++i) {
-        offsetSum += setFileOffset((*targetList)[i], offsetSum);
+        offsetSum += setFileOffset((*targetList)[i], (size_t) offsetSum);
     }
     return offsetSum - fileOffset;
 }
